% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_npp_model}
\alias{load_npp_model}
\title{Load the pre-trained NPP prediction model}
\usage{
load_npp_model()
}
\value{
A random forest model object of class "train" from the caret package.
  The object contains the fitted model with 1000 trees.
}
\description{
Loads the random forest model included with the package. The model was
trained on a global compilation of brGDGT data and is used by 
\code{\link{predict_npp}}.
}
\examples{
\dontrun{
# Load the global model
model <- load_npp_model()

# Examine model structure
print(model)

# Get model summary
summary(model)
}

}
\seealso{
\code{\link{predict_npp}} for making predictions,
\code{\link{diagnose_applicability}} for checking data suitability
}
