% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_model.R
\name{print.brGDGT_diagnostic}
\alias{print.brGDGT_diagnostic}
\title{Print method for diagnostic objects}
\usage{
\method{print}{brGDGT_diagnostic}(x, ...)
}
\arguments{
\item{x}{brGDGT_diagnostic object}

\item{...}{Additional arguments}
}
\description{
Print method for diagnostic objects
}
